New Feature - Zoho Gateway
Zoho Webhooks are not data-compatible with Mattermost, so we need a
gateway to reformat what we get.

This adds a generic webhook with a configurable URL. You can then POST
anything to it with a message parameter (POST data or query string is
accepted). This will be sent to the configured mattermost channel.